chore: Rename databasefake --> dbfake
More consistent naming
