+++ coderd/database/dbfake/databasefake.go
@@ -1,4 +1,4 @@
-package databasefake
+package dbfake
 
 import (
 	"context"
