+++ enterprise/audit/backends/postgres_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/coder/coder/coderd/database"
-	"github.com/coder/coder/coderd/database/databasefake"
+	"github.com/coder/coder/coderd/database/dbfake"
 	"github.com/coder/coder/enterprise/audit/audittest"
 	"github.com/coder/coder/enterprise/audit/backends"
 )
@@ -19,7 +19,7 @@ func TestPostgresBackend(t *testing.T) {
 
 		var (
 			ctx, cancel = context.WithCancel(context.Background())
-			db          = databasefake.New()
+			db          = dbfake.New()
 			pgb         = backends.NewPostgres(db, true)
 			alog        = audittest.RandomLog()
 		)
